lambdas/resize/requirements.txt
boto3
Pillow==9.5.0